Fix Snake.toString method bug introduced with previous change

+++ server/src/main/java/game/snake/Snake.java
@@ -298,6 +298,6 @@ public class Snake {
 
     @Override
     public String toString() {
-        return "Snake " + id;
+        return "Snake " + ((int) id);
     }
 }
